backend/requirements.txt
# Core dependencies
boto3>=1.34.0
anthropic>=0.40.0
pyyaml>=6.0.0
pydantic>=2.0.0

# Development dependencies (not needed in Lambda)
# pytest>=8.0.0
# pytest-cov>=4.1.0
# moto>=5.0.0
